refactor: 린트 체크 및 패키지 수정
