+++ app/src/main/java/com/teamwss/websoso/ui/common/base/BindingActivity.kt
@@ -1,4 +1,4 @@
-package com.created.team201.presentation.common
+package com.teamwss.websoso.ui.common.base
 
 import android.os.Bundle
 import androidx.annotation.LayoutRes
