+++ app/src/main/java/com/teamwss/websoso/ui/explore/ExploreFragment.kt
@@ -1,8 +1,8 @@
 package com.teamwss.websoso.ui.explore
 
-import com.created.team201.presentation.common.BindingFragment
 import com.teamwss.websoso.R
 import com.teamwss.websoso.databinding.FragmentExploreBinding
+import com.teamwss.websoso.ui.common.base.BindingFragment
 
 
 class ExploreFragment : BindingFragment<FragmentExploreBinding>(R.layout.fragment_explore)
