+++ app/src/main/java/com/teamwss/websoso/ui/feed/FeedFragment.kt
@@ -1,7 +1,7 @@
 package com.teamwss.websoso.ui.feed
 
-import com.created.team201.presentation.common.BindingFragment
 import com.teamwss.websoso.R
 import com.teamwss.websoso.databinding.FragmentFeedBinding
+import com.teamwss.websoso.ui.common.base.BindingFragment
 
 class FeedFragment : BindingFragment<FragmentFeedBinding>(R.layout.fragment_feed)
