+++ app/src/main/java/com/teamwss/websoso/ui/home/HomeFragment.kt
@@ -1,7 +1,7 @@
 package com.teamwss.websoso.ui.home
 
-import com.created.team201.presentation.common.BindingFragment
 import com.teamwss.websoso.R
 import com.teamwss.websoso.databinding.FragmentHomeBinding
+import com.teamwss.websoso.ui.common.base.BindingFragment
 
 class HomeFragment : BindingFragment<FragmentHomeBinding>(R.layout.fragment_home)
