+++ app/src/main/java/com/teamwss/websoso/ui/main/MainActivity.kt
@@ -6,12 +6,13 @@ import androidx.annotation.IntegerRes
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.commit
 import androidx.fragment.app.replace
-import com.created.team201.presentation.common.BindingActivity
 import com.teamwss.websoso.R
 import com.teamwss.websoso.databinding.ActivityMainBinding
+import com.teamwss.websoso.ui.common.base.BindingActivity
 import com.teamwss.websoso.ui.explore.ExploreFragment
 import com.teamwss.websoso.ui.feed.FeedFragment
 import com.teamwss.websoso.ui.home.HomeFragment
+
 import com.teamwss.websoso.ui.main.MainActivity.FragmentType.EXPLORE
 import com.teamwss.websoso.ui.main.MainActivity.FragmentType.FEED
 import com.teamwss.websoso.ui.main.MainActivity.FragmentType.HOME
