+++ app/src/main/java/com/teamwss/websoso/ui/myPage/MyPageFragment.kt
@@ -1,7 +1,7 @@
 package com.teamwss.websoso.ui.myPage
 
-import com.created.team201.presentation.common.BindingFragment
 import com.teamwss.websoso.R
 import com.teamwss.websoso.databinding.FragmentMyPageBinding
+import com.teamwss.websoso.ui.common.base.BindingFragment
 
 class MyPageFragment : BindingFragment<FragmentMyPageBinding>(R.layout.fragment_my_page)
