+++ app/src/main/res/drawable/ic_explore.xml
@@ -3,10 +3,10 @@
     android:height="22dp"
     android:viewportWidth="23"
     android:viewportHeight="22">
-  <path
-      android:pathData="M1.668,5.2C1.668,4.08 1.668,3.52 1.886,3.092C2.078,2.716 2.384,2.41 2.76,2.218C3.188,2 3.748,2 4.868,2H5.268C7.508,2 8.628,2 9.484,2.436C10.237,2.819 10.849,3.431 11.232,4.184C11.668,5.04 11.668,6.16 11.668,8.4V20L11.568,19.85C10.873,18.808 10.526,18.287 10.067,17.91C9.661,17.576 9.193,17.325 8.69,17.173C8.121,17 7.495,17 6.243,17H4.868C3.748,17 3.188,17 2.76,16.782C2.384,16.59 2.078,16.284 1.886,15.908C1.668,15.48 1.668,14.92 1.668,13.8V5.2Z"
-      android:fillColor="#AEADB3"/>
-  <path
-      android:pathData="M21.668,5.2C21.668,4.08 21.668,3.52 21.45,3.092C21.258,2.716 20.952,2.41 20.576,2.218C20.148,2 19.588,2 18.468,2H18.068C15.828,2 14.708,2 13.852,2.436C13.099,2.819 12.487,3.431 12.104,4.184C11.668,5.04 11.668,6.16 11.668,8.4V20L11.768,19.85C12.463,18.808 12.81,18.287 13.269,17.91C13.675,17.576 14.143,17.325 14.646,17.173C15.215,17 15.841,17 17.093,17H18.468C19.588,17 20.148,17 20.576,16.782C20.952,16.59 21.258,16.284 21.45,15.908C21.668,15.48 21.668,14.92 21.668,13.8V5.2Z"
-      android:fillColor="#AEADB3"/>
+    <path
+        android:fillColor="#AEADB3"
+        android:pathData="M1.668,5.2C1.668,4.08 1.668,3.52 1.886,3.092C2.078,2.716 2.384,2.41 2.76,2.218C3.188,2 3.748,2 4.868,2H5.268C7.508,2 8.628,2 9.484,2.436C10.237,2.819 10.849,3.431 11.232,4.184C11.668,5.04 11.668,6.16 11.668,8.4V20L11.568,19.85C10.873,18.808 10.526,18.287 10.067,17.91C9.661,17.576 9.193,17.325 8.69,17.173C8.121,17 7.495,17 6.243,17H4.868C3.748,17 3.188,17 2.76,16.782C2.384,16.59 2.078,16.284 1.886,15.908C1.668,15.48 1.668,14.92 1.668,13.8V5.2Z" />
+    <path
+        android:fillColor="#AEADB3"
+        android:pathData="M21.668,5.2C21.668,4.08 21.668,3.52 21.45,3.092C21.258,2.716 20.952,2.41 20.576,2.218C20.148,2 19.588,2 18.468,2H18.068C15.828,2 14.708,2 13.852,2.436C13.099,2.819 12.487,3.431 12.104,4.184C11.668,5.04 11.668,6.16 11.668,8.4V20L11.768,19.85C12.463,18.808 12.81,18.287 13.269,17.91C13.675,17.576 14.143,17.325 14.646,17.173C15.215,17 15.841,17 17.093,17H18.468C19.588,17 20.148,17 20.576,16.782C20.952,16.59 21.258,16.284 21.45,15.908C21.668,15.48 21.668,14.92 21.668,13.8V5.2Z" />
 </vector>
