+++ app/src/main/res/drawable/ic_home.xml
@@ -3,8 +3,8 @@
     android:height="18dp"
     android:viewportWidth="18"
     android:viewportHeight="18">
-  <path
-      android:pathData="M7.891,0.738C8.562,0.29 9.438,0.29 10.109,0.738L17.109,5.405C17.666,5.776 18,6.4 18,7.069V15.999C18,17.103 17.105,17.999 16,17.999H2C0.895,17.999 0,17.103 0,15.999V7.069C0,6.4 0.334,5.776 0.891,5.405L7.891,0.738Z"
-      android:fillColor="#AEADB3"
-      android:fillType="evenOdd"/>
+    <path
+        android:fillColor="#AEADB3"
+        android:fillType="evenOdd"
+        android:pathData="M7.891,0.738C8.562,0.29 9.438,0.29 10.109,0.738L17.109,5.405C17.666,5.776 18,6.4 18,7.069V15.999C18,17.103 17.105,17.999 16,17.999H2C0.895,17.999 0,17.103 0,15.999V7.069C0,6.4 0.334,5.776 0.891,5.405L7.891,0.738Z" />
 </vector>
